Move all effect runners to a separated file
