+++ packages/core/src/internal/runSaga.js
@@ -1,7 +1,8 @@
 import * as is from '@redux-saga/is'
 import { compose } from 'redux'
-import { check, uid as nextSagaId, wrapSagaDispatch, noop, log as _log, identity } from './utils'
-import proc, { getMetaInfo } from './proc'
+import { check, log as _log, noop, uid as nextSagaId, wrapSagaDispatch, identity } from './utils'
+import { getMetaInfo } from './error-utils'
+import proc from './proc'
 import { stdChannel } from './channel'
 import { immediately } from './scheduler'
